Updated openstack/openstack Project: openstack/neutron  af0e0edce9483785646b196f071087b343ea59b0
Validate device to mac instead of port id to mac

When updating the Fdb table, validate if rule exists with
device to mac instead of port id to mac.
In case several devices are located on the same physnet each device
needs to be updated separately, therefor the validation of existing
Fdb rules should be device to mac.

Change-Id: I889cbb02a875403122d520357c38eae2af14ebbe
Closes-Bug: #1604838