Updated openstack/openstack Project: openstack/cinder  fa8201b7ce7ddbcf5480fbaf154f9eb92be35d62
Add lock decorator to SolidFire clone_image method

When using the SolidFire template-caching feature, if a
template volume doesn't exist and multiple calls are made
at the same time to create a volume from image, we can run
in to a couple of problems.

1. It's possible to clone the template volume before the
   image conversion is actually completed (so the clone is
   a corrupt volume).
2. Hit a race where a subsequent call doesn't get the in
   progress template create and creates a duplicate.

We can fix this by using a lock decorator around the
clone_image method in the driver.

Change-Id: I0000c5bf29629fc1ded7b9e3ae26989fbf9dc286
Closes-Bug: #1605027