Updated openstack/openstack Project: openstack/os-net-config  f17add2e8e3c40f79cf211b0cb82a359104ad675
Add support for Infiniband interfaces

This patch adds support for Infiniband interfaces. The only difference
between Inifiniband and regular interfaces at this time is that an
interface with type "ib_interface" will have "TYPE=Infiniband" added
to the ifcfg file.

However, the Infiniband interface is implemented as a full new class,
so in the future we can add script functions or additional config
options to the Infiniband interface config if needed.

Unit tests for both the object and the ifcfg implementation are
included. This patch does not include an implementation for systems
that use /etc/network/interfaces (Debian-based systems).

Note that this change has not yet been tested on bare metal with
Infiniband hardware.

Fixes bug: https://bugzilla.redhat.com/show_bug.cgi?id=1326616

Change-Id: Iaeaca9cd71e2cea6147951d49aecc7458be4ca0b