Updated openstack/openstack Project: openstack/nova  5c70014e13272c0b41a3c4b510217d2ef4f46652
Warn when using null cache backend

The default backend for oslo_cache is dogpile.cache.null, leading to a
broken setup when following the current release notes or the deprecation
message verbatim. So we should at least emit a warning when the config
does not specify a different backend.

Note: I'm not sure whether it is possible to amend the release note like
this or whether there needs a new note to be added, please advise.

Change-Id: I16647e424d8382dae98f13cb1f73a7e0c0aebaf5
Closes-Bug: 1572062