Updated openstack/openstack Project: openstack/nova  08c9481ed86042f93fe954ebc80fd61a6e63d2f5
network: handle unauthorized exception from neutron

Neutron can raise an unauthorized exception for an expired or invalid
token. In case of admin context we log a message to inform operators
that the Neutron admin creadential are not well configured then return
a 500 to user. In case of user context we return 401.

Change-Id: I87c8b86373967639eb55b4cc3b7d6cbd9780f3ac
Closes-Bug: #1571722