Updated openstack/openstack Project: openstack/nova  a0bbc80aa7af65400a80eb6da565e8a5ce1b3223
Clean up test_check_attach_availability_zone_differs

This test is really four tests, two of which don't make
any difference and one is a happy path test which is
redundant with the test_check_attach test that comes right
after it.

The two tests which don't make any difference are the ones
where the instance.host is set to None. This made a difference
before az.get_instance_availability_zone was mocked out but
it's not used anymore in this test.

There are some other cleanups in this change:

- move the mock to be a decorator
- use an instance object
- use a valid uuid for the volume id
- use self.flags so we don't have to reset the config override

Change-Id: I44f32f10d99f7cc6edce1d5ae4bc359c6d14fdf3