Updated openstack/openstack Project: openstack/nova  4a772b1fc5e75341c11efe0e9195670c1371fd56
Make Aggregate host operations work against API db

Adds the ability to add and delete hosts to Aggregates that are
contained in the API db. Modifications will be made to the
'aggregate_hosts' table in the API db if that is where the aggregate is
located, or the cell database otherwise.

blueprint cells-aggregate-api-db

Change-Id: Icb11e5cbcc743d1ae788c792cffba9de00fddc3f