Updated openstack/openstack Project: openstack/keystone  84c0c74dd9548f8ba1da8f8e766eccc4c9cd4601
Refactor TestAuthExternalDomain to not inherit tests

Previously, TestAuthExternalDomain was inheriting from test_v3.RestfulTestCase,
which allowed it to run as part of the keystone test suite. This commit breaks
it into a class that only inherits from `object` and introduces 3 other classes
the inherit the old TestAuthExternalDomain and run the tests according to the
setup needed.

Since the Fernet provider doesn't support bind authentication, there is no test
class to setup Fernet and run the TestAuthExternalDomain behaviors. This change
will make defaulting to Fernet easier.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: I28e575ddada8492bd4fc17b78cb00651d9d4af07
Partial-Bug: 1561054