Updated openstack/openstack Project: openstack/keystone  f47fe6247d07e8812856a8b8e948cbf8895f96ef
Don't run TokenCacheInvalidation with Fernet

Since the TokenCacheInvalidation tests are specific to token formats that
require persistent storage, it doesn't make sense to run them with Fernet
configured. This commit updated the classes the inherit TokenCacheInvalidation
to test all token formats except the Fernet format.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: Iaaf0f29fa1d55b0028729b69704167957a6c5f8c
Partial-Bug: 1561054