Updated openstack/openstack Project: openstack/keystone  663ff28d604f48d8b8086a84071b9c3e26020216
refactor: make TestAuthKerberos test pki/pkiz/uuid

This helps us move towards fernet because we don't support bind authentication
with the Fernet provider, so when we set Fernet to be the default we won't run
these tests. We will only run Kerberos tests against token providers that
support it.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: Ie80f74e47046c5d8d3c2f31e5b84e4210e775f7f
Partial-Bug: 1561054