Updated openstack/openstack Project: openstack/keystone  8b47d17ed778f32deb925c1abbd146bd990779b1
refactor: make TestAuthExternalDefaultDomain test uuid/pki/pkiz

Bind authentication is only supported by uuid, pki, and pkiz token providers.
This makes it easier to switch Fernet to be the default.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: I195ae1def4e2e7e27125f09cf058718b73eb839e
Partial-Bug: 1561054