Updated openstack/openstack Project: openstack/keystone  321d7a8a4234b3edc23c88edb4028f7289902cb1
Use freezegun in OSRevokeTests

This makes it easier for us to switch fernet to be the default token provider
because we can increment the clock in order to get the list of revocation
events.

This is an effort to break https://review.openstack.org/#/c/258650 into
smaller, more reviewable pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: Ia47a78509d033596b0101b28e9cd38abafdb199a
Partial-Bug: 1561054