Updated openstack/openstack Project: openstack/ceilometer  1e603527696c40b316a6dfc73d87433644bf54d0
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I3c4e7de5a121d7c5d5a7a0f8bba5cb6014da7afc
Closes-bug: #1604734