Updated openstack/openstack Project: openstack/keystone  89c160dd818c5d65664f51174f39fbc216f25ff3
Use freezegun in AssignmentInheritanceTestCase

This makes it so that we properly advance the clock after revocation events.
Is also makes it easier to switch to Fernet as the default token provider.

This fix was originally a part of https://review.openstack.org/#/c/258650 but
this is an attempt to break 258650 into smaller, more reviewable, pieces.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Adam Young <ayound@redhat.com>

Change-Id: I83509e81e4da26cad48ed3cb3f9733ddb7ae965e
Partial-Bug: 1561054