Updated openstack/openstack Project: openstack-dev/devstack  b12c25a3e60b6deeceec54ccca1592b20bfdb949
Keep old behavior of setuptools for editable installs

In the 25.0.0 release [1] of setuptools during any install
operation the package in not overwritten. If a package is
installed from another requirement via pip and then it is
installed again from git, it is not updated causing
check_libs_from_git to fail.

[1] https://setuptools.readthedocs.io/en/latest/history.html#v25-0-0

Change-Id: Ibaa1d4157816ea649f4452756fbde25951347001
Closes-Bug: #1605998