Updated openstack/openstack Project: openstack/keystone  96af93d36151b04fc82c5c1db3615311bfda544e
Fix v2-ext API enabled documentation

The 'enabled' key in the parameters list refers to a tenant, not a
user. The `/v2.0/users/{userId}/OS-KSADM/enabled` method is operating
on a user, so we should use the 'enabled_1' key. 'enabled_1' is not a
very descriptive name so let's call it 'enabled_user' instead.

Moreover, this parameter is required, so it does not make sense for it
to have a default when the user must supply a value, so this patch
modifies the 'enabled_1' (now called 'enabled_user') description to
reflect that.

Change-Id: I81a7c68d5000ca6acab74679b49eeb619a5b914e