Updated openstack/openstack Project: openstack/nova  7d505bd82014377aa081935bd613628ed87b2162
Deprecate network quota configuration

The quota_networks option is only used for the os-tenant-networks
API which only supports CRUD operations for nova-network which
is deprecated, so this change deprecates the quota_networks option.

Furthermore, tenant network quota is disabled by default due to
the enable_network_quota option, so we also deprecate that option
here.

Change-Id: I892855ccf8ae83c0bd5e2bc25ea64dee83ab79ed