Updated openstack/openstack Project: openstack/heat  9f0c2651e8435cc6e78da3f11bfeb077d9759e4c
Don't persist_state of IN_PROGRESS stack again

When a greenthread is stopped(ex. deleting an IN_PROGRESS stack),
the linked function(release) is called and it tries to save
the state of the stack. As we already saved the state of the
IN_PROGRESS stack, we should not try to do it again.

Change-Id: I598afcd286881bc1d9394152fe38701d5fed473f
Partial-Bug: #1546431