Updated openstack/openstack Project: openstack/openstack-manuals  4a6d8b05f71d6344286ffdbee8c17c5fed707698
Adds the UX Personas descriptions.

This User Experience personas are meant to help developers
identify the target audience for their work.
This is the first pass for the content.
The intent is to provide the general information for all five personas
in a consistent manner.

Additional content such as infographics, comparison tables, etc., will
be part of subsequent additions.

Partial-bug: #1603420

Change-Id: I8fa4e0de762b97f44137c2ea48c0a9ee5b8bc61a
Implements: blueprint ux-personas
Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>