Updated openstack/openstack Project: openstack/python-keystoneclient  09f47ab0ba9b0b7a7b46dbb6d60ad7f63dff7da7
Use assertEqual() instead of assertDictEqual()

OpenStack now has dropped support for python 2.6 and lower.
So we don't need to define an assertDictEqual() for python
lower than 2.7.

Further, assertEqual() in testtools is able to handle dicts
comparation, just the same as assertDictEqual() in unittest2.
So we do't need to call assertDictEqual() for dicts any more.

Please also refer to: https://review.openstack.org/#/c/347097/

Change-Id: Ieaf211617c38aa0f9a38625b1009c36bd6a16fba