Updated openstack/openstack Project: openstack/nova  1913050fd1eafad41e0e8f6f8f9795bf848ed5d2
raise exception ComputeHostNotFound if host is not found

When performing a live migration,I passed a wrong host parameter,
and then I got a wrong result like this:
"Compute service of host-1 is unavailable at this time".
I thought the compute service stopped, but the fact was I passed
a host id instead of a host name,therefore it could not find the host.

In order to distinguish the error "host not found" from the error
"service not available", I think we should raise a different exception
ComputeHostNotFound instead of ComputeServiceUnavailable.

This patch we will return more accurate error msg for live migration.

Closes-Bug: #1538837
Change-Id: I6ad377147070f85b9b1d5a1d1be459e890e02bcc