Updated openstack/openstack Project: openstack/ironic  f7bfa3ee73a2f94be28017212b699d07e9c7ad92
Remove unused code when failing to start console

Processutils.ProcessExecutionError is never raised when it is unable
to stop an old shellinabox process due to a fix[1][2]. Instead of the
error, ConsoleError is raised. It would be better not to ignore the
error and to fail to start a console for avoiding process leak.

This patch removes a block for the exception which is never raised.

[1]https://review.openstack.org/#/c/242319/
[2]https://review.openstack.org/#/c/244699/

Change-Id: Ia1226da652510c5655c9b8c270e6fc168efd935d