Updated openstack/openstack Project: openstack/tempest  c9436c4fffe403be3555d1738440355019fffb35
Introduce the ClientsFactory

Define a new class ClientsFactory, which provides lazy loading
an simple instantiation of clients by using partial to set
defaults for all client parameters.

This class will be used in upcoming patches to implement an
interface that allows plugins to register their own service
clients in the ServiceClients interface.

It will also be used to implement lazy loading of clients for
tempest own clients.

Partially-implements: bp client-manager-refactor

Change-Id: If2295d56a8e52ffce28d6a0f15517bc325aab010