Updated openstack/openstack Project: openstack-infra/tripleo-ci  e3e27abeb011497bbadf736c4f2d5caa1a621a3c
Remove sahara and cinder packaged symlinks in postci

If a job fails before the overcloud deployment starts, the ansible task
run by infra to collect logs will fail because of dangling symlinks on
the subnodes in the /etc/cinder and /etc/sahara directories.

This change just remove the symlinks on each subnodo so that the rsync
command to collect the logs won't fail.

Change-Id: If0a151abdeaa1e607d672175f8bcb170027b9ac6
Closes-Bug: #1606368