Updated openstack/openstack Project: openstack/cinder  eabd2e87184dda4d1a14b13e306ac4edf284eaf8
Update doc README and remove old Makefile

The information in the README file in the doc directory was very out of
date and the Makefile referenced in it doesn't appear to work any more.

Our preferred doc build method is to stay consistent with using tox -e docs,
so removing the legacy build information and providing better details for
doc development.

Change-Id: Ifd01b9cfc39fe99697a1669ecb5db6302359b662