Updated openstack/openstack Project: openstack/cinder  c523afa0fdbc6c46270419fbdb71c8bc6103ed52
Move fake_driver to tests root dir

The fake driver is now being used for both unit and
functional tests.  It also now has a FakeGate driver to
add some things that aren't supported by all drivers
(including the reference as well as ceph and others).

This patch just moves the fake_driver.py file from
tests/unit to tests as it's not specifically a unit test
object.  This isn't a big deal, but makes things a bit more
clear and obvious that there's a useful fake that can be
used in a number of places.

Some of the other fakes could likely be moved as well, but
this one is a start.

Change-Id: Ia5e8608832c0753a4ced8690fa285184c5b8c0c1