Updated openstack/openstack Project: openstack/neutron  a0936bb3f0aa456e0504dd7da4123a7f8bc7f4f7
Prevent port update from binding a host where IPs won't work

In the context of a routed network, if a port has fixed ips already
then those fixed ips can only be used on the segment with that subnet
on it. If a port update tries to update the host binding to a
different segment, an exception will now be raised.

Change-Id: I8dc8890907d1e241dd12448fa184cea1b0620663
Partially-Implements: blueprint routed-networks