Updated openstack/openstack Project: openstack/glance_store  10f50da5fcd9257090ab137f4d257940bb01b44a
Fix cinder config string as per current i18n state

As described in the bug 1605648 , the translations being done on the
config string are not permitting additions so, the glance docs
generation is failing.

The glance_store 0.14.0 had to be pinned down to let glance review queue
not break. see updates on this thread
http://lists.openstack.org/pipermail/openstack-docs/2016-July/thread.html#8899

The config strings need to be non additive for translations and will
enable future releases of glance_store. Once, a appropriate fix is
published for bug 1605648 the config strings can be more flexible.

Change-Id: Ia6ed5883458f2e81e861cda02068956728e76357
Closes-Bug: 1606746
Partial-Bug: 1605648