Updated openstack/openstack Project: openstack/nova  356b44c2aa1ac0b746414add06ccf00e82a3ac84
Remove deprecated ironic.api_version config option

The option was deprecated in mitaka:

26c69a6a62a138ccdb88ecc5bf75330a04e5de0b

Some parts of the code were still using it, but it's
effectively not used since that change in mitaka and
we can just hard-code it when getting the ironic client,
which also doesn't use it:

http://git.openstack.org/cgit/openstack/python-ironicclient/tree/ironicclient/client.py?h=1.1.0#n58

There is no release note for this change. Since the option was
made useless in Mitaka there isn't really a point in notifiying
people that a vestigial config option is now gone.

Change-Id: If594ae1a367b20db9b2d65188d65a0c6ab09789b