Updated openstack/openstack Project: openstack/neutron  15196c42803e2d78e046a2f9a51a43a6c55c167f
tests: stop using ml2 plugin full import paths in tests

Stevedore documentation suggest that full import paths are not supposed
to be user visible. Since unit tests emulate users when configuring
oslo.config, we better off relying on well known plugin aliases than
internal details.

For in-tree that may be not a big deal, but with it we set a bad example
for third parties that may later find their tests broken eg. when we
decide to move code around.

TrivialFix

Change-Id: I7bd036ac3df7e7f4c678356d0a793e7d38599dda