Updated openstack/openstack Project: openstack/oslo.config  b40925328e5ba9e87be42bab1ec3a41bf3d81f78
disable lazy translation in sphinx extension

The sphinx extension depends on all of the help text associated with an
option being regular strings. When oslo_i18n's lazy translation feature
is turned on, the marker functions return Message objects that do not
support all of the features of strings. Add a step to turn off lazy
evaluation when the oslo_context.sphinxext extension is activated.

Change-Id: Icaec4691135cd07d96994f01498e0f9d2e09af58
Closes-Bug: #1605648
Signed-off-by: Doug Hellmann <doug@doughellmann.com>