Updated openstack/openstack Project: openstack/oslo.utils  f43d78d358aebba94b8cf644b1352a8b9e401f53
Extend specs matcher to support ">" and "<"

This patch is exteding the specs matcher and adding support for the ">"
and "<" operators. Prior to this patch, only the "equal" forms were
supported (">=" and "<=").

It also looks inconsistent because for the "s" prefixed operators all
the 4 forms are supported already: "s<=", "s<", "s>=" and "s>".

Change-Id: I82a72b0ef6ee277e7b09b0eb138e058687ce8804