Updated openstack/openstack Project: openstack-infra/puppet-unbound  dffe3580c13ee7519f7e119a31c9e7acc31c99e7
Update dhclient.conf to be in sync across all OSes

After finding some puppet jobs were failing on CentOS it was found the
DHCP server for osic and ovh were passing extra settings.

As a results, we also noticed our dhclient.conf was not in sync around
all of our images.  So, this patch attemps to fix both issues.

Change-Id: I2415f7e42a9d5931589192b7d78423941e02b076
Signed-off-by: Paul Belanger <pabelanger@redhat.com>