Updated openstack/openstack Project: openstack/manila  60a32f13f8880d19fda13c78af4403e58f1a8fc6
NetApp: Report hybrid aggregates in share stats

NetApp cDOT controllers can mix SSDs and spinning disks
in the same aggregate, where the SSDs are used for a
cache.  This commit reports the hybrid aggregate attribute
as well as the aggregate name to the scheduler for each
pool for use in extra specs matching.

Implements: blueprint netapp-report-cdot-hybrid-aggrs-manila
Change-Id: Iaa0bcd79789449f977b48f1de2adf997c936db61