Updated openstack/openstack Project: openstack/python-ironicclient  98c4f4dbeae22536b6f5fd83c6cec815fa2fa42f
Fix coverage target

Previously the coverage target ended up deleting the coverage data when
the `coverage combine` command ran. This then caused the `coverage html`
command to fail due to lack of data to report on.

Thankfully the fix is simple because pbr is actually handlign all of
this for us behind the scenes. All we need to do is run with the
--coverage flag and make sure the .coveragerc reflects the desired
inclusions and ommisions.

Change-Id: Ide6c7639a19d961e4af27c2c59b036395e2af3ef