Updated openstack/openstack Project: openstack/nova  3645e55ee3148993f8f6c1abfda6a6f9e345bb3d
Call release_dhcp via RPC to ensure correct host

When deleting an instance in a nova-network environment, the network
manager calls release_dhcp() on the local host. The linux_net driver
then executes dhcp_release, a binary that comes with dnsmasq that
releases a DHCP lease on the local host. Upon lease release, dnsmasq
calls its dhcp-script, nova-dhcpbridge. The latter calls
release_fixed_ip() and the instance's fixed IP is returned to the
pool. This is fine if an instance has never been live-migrated.

If an instance has been live-migrated, the dnsmasq on its new
host fails with 'unknown lease' because it's not the same dnsmasq that
originally handed out the lease. Having failed, dnsmasq doesn't call
nova-dhcpbridge and release_fixed_ip() is never called. The fixed IP
is not returned to the pool and a new instance cannot be booted with
that IP.

This patches adds a release_dhcp RPC call that calls release_dhcp on
the instance's "original" host, thus ensuring that the correct dnsmasq
handles the lease release and that nova-dhcpbridge and
release_fixed_ip() are called.

Change-Id: I0eec8c995dd8cff50c37af83018697fc686fe727
Closes-bug: 1585601