Updated openstack/openstack Project: openstack-infra/project-config  e341966c776aa8006318a8e51822f5d7bbca0602
add gentoo support to bindep-fallback.txt

This commit adds gentoo support for packages listed in
bindep-fallback.txt.  It should be all of them except zookeeperd (we
don't have that packaged).  A few of these are system things that are
installed by default.  This should be a good starting point we can
work from though.

Change-Id: Ib3e3471d70b963308cc58d30d12fbdd5ff493438