Updated openstack/openstack Project: openstack-infra/tripleo-ci  b17eefca280f0e8c45f3fda46bf351f73139a8f6
Fix CI status report to show all job results on 1 line

When we added the multinode job to the CI status report, we introduced a
bug whereby the job results for the same patchset were split across
multiple lines in the report.

The bug was due to the fact that the script assumes that all of the CI
results were in a single comment from Jenkins in the review. With the
addition of the multinode job, this is no longer true: we have to parse
multiple comments.

The patch changes the building of the results dictionary to update it's
data (using setdefault) to support parsing of multiple comments and keys
by review id and patchset instead of timestamp.

The ci_results dict in each value of the results dict should now be a
union of the CI results from all the latest comments that contain the
job names we are checking.

During the html generation, we still want to sort by timestamp, so we
need to introspect the values of the "data" dict instead of just sorting
by its keys.

Change-Id: I8df41877ea274170f7466bf35cb040c46dea58d2