Updated openstack/openstack Project: openstack/neutron  ed08c60617faa3dd6ac4f8e26263e12256321899
Pass timeout in milliseconds to timer_wait

The OVS Python library expects milliseconds to be passed to
timer_wait and we were passing the value in seconds. This
just means that the poll() wakes up more often than it needs
to. Changing this probably improves performance a little, but
doesn't change any user observable behavior.

Change-Id: Iae4b682c0774e749eb48cde67cd9748d7dfe053c