Updated openstack/openstack Project: openstack-infra/project-config  bb14fadb7d3b95fb1a9917ca28d08dbedb8638ad
Tighten the trusty job filter for OpenStack-Ansible properly

Only the functional jobs should run, not the docs jobs too. As the
functional jobs could be 'func' or any other name, a negative match
is used on the docs job name instead.

The previous attempt at this didn't work properly.

The following should not pass the regex test:
- gate-openstack-ansible-docs-ubuntu-trusty-nv
- gate-openstack-ansible-docs-ubuntu-trusty
- gate-openstack-ansible-repo_build-ansible-docs-ubuntu-trusty-nv
- gate-openstack-ansible-repo_build-ansible-docs-ubuntu-trusty

The following should pass the regex test:
- gate-openstack-ansible-func-ubuntu-trusty-nv
- gate-openstack-ansible-func-ubuntu-trusty
- gate-openstack-ansible-repo_build-ansible-func-ubuntu-trusty-nv
- gate-openstack-ansible-repo_build-ansible-func-ubuntu-trusty

Change-Id: I595c926d1e441777303c6f5ecb68ba72abbae0c6