Updated openstack/openstack Project: openstack/keystone  d07f59ff38e3993d288b71d55436e507366cf98c
Invalid tls_req_cert constant as default

There are several occurances where the argument default of
tls_req_cert is set to 'demand'. This is an invalid option for the
function _common_ldap_initialization, which expects one of the values
of the dict LDAP_TLS_CERTS, not a key.

If tls_req_cert='demand' is given to _common_ldap_initialization,
you'll get:
   "LDAP TLS: invalid TLS_REQUIRE_CERT Option=demand"

Change-Id: I2da2c041bac25ca7ebab8c32b464164279a6bd69